% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acs_data.R
\name{acsdata}
\alias{acsdata}
\title{Create a list containing ACS data}
\usage{
acsdata(formula, state = "WI", level = "county", endyear = 2013,
  county = "*", county_subdivision = "*", tract = "*",
  block_group = "*", congressional_district = "*",
  school_district_secondary = "*", school_district_elementary = "*")
}
\arguments{
\item{formula}{A vector or character object following the format.}

\item{state}{A number or character specifying state.}

\item{level}{A character or vector specifying the}
}
\value{
List of ACS objects of different levels.
}
\description{
Create a list containing ACS data
}
\examples{
add(1, 1)
add(10, 1)
}

