% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acsr.R
\docType{package}
\name{acsr}
\alias{acsr}
\alias{acsr-package}
\title{acsr: A package with wrapped functions to manage ACS files.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

