% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percent_change_field.R
\name{calculate_percent_change_field}
\alias{calculate_percent_change_field}
\title{Calculate change in a variable from one year to another}
\usage{
calculate_percent_change_field(
  con,
  variable_name,
  variable,
  derived_status,
  from_year,
  to_year,
  geography,
  keys,
  params
)
}
\arguments{
\item{con}{A database connection}

\item{variable_name}{Name of the variable to be derived}

\item{variable}{The ACS variable to calculate change in}

\item{derived_status}{If TRUE, use MDA derived variable to calculate change}

\item{from_year}{Base year for comparison}

\item{to_year}{End year to calculate change against}

\item{geography}{Geographic level of data}

\item{keys}{Named list with the structure `list(geoid = X, year = 'year')`, where X is variable name for the GEOID of the geography}

\item{params}{List of params}
}
\value{
A data frame
}
\description{
Calculate change in a variable from one year to another
}
