% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage_field.R
\name{calculate_percentage_field}
\alias{calculate_percentage_field}
\title{Calculate a percentage from ACS fields}
\usage{
calculate_percentage_field(
  con,
  variable_name,
  numerator_cols,
  denominator_cols,
  inverse_pct,
  geography,
  keys,
  params
)
}
\arguments{
\item{con}{A database connection}

\item{variable_name}{Name of the variable to derive}

\item{numerator_cols}{Variable or variables to divide. Multiple values will be summed.}

\item{denominator_cols}{Variable or variables to divide by. Multiple values will be summed.}

\item{inverse_pct}{If TRUE, result will be 1 - (sum(numerator_cols)/sum(denominator_cols))}

\item{geography}{Geographic level of data}

\item{keys}{Named list with the structure `list(geoid = X, year = 'year')`, where X is variable name for the GEOID of the geography}

\item{params}{List of params}
}
\value{
A data frame
}
\description{
Calculate a percentage from ACS fields
}
