% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_make_geo.R
\name{combine.make.geo}
\alias{combine.make.geo}
\title{Combine geographic sets using the ACS geo.make function}
\usage{
combine.make.geo(level, combine.name = "aggregate", print.levels = TRUE,
  region = "*", division = "*", state = "*", county = "*",
  county.subdivision = "*", place = "*", tract = "*", block.group = "*",
  msa = "*", csa = "*", necta = "*", urban.area = "*",
  congressional.district = "*", state.legislative.district.upper = "*",
  state.legislative.district.lower = "*", puma = "*", zip.code = "*",
  american.indian.area = "*", school.district.elementary = "*",
  school.district.secondary = "*", school.district.unified = "*")
}
\arguments{
\item{level}{A character or vector of characters specifying the geographic
  level of the data. It may be necessary to specificy values to the
  corresponding levels. For instance, when \code{level = "county"}, you have
  to specify a state (e.g., \code{state = "WI"}, the default state in this
  package). You can also use a wildcard method (\code{state = "*"}) to
  include all the states. Below, you can see the required combinations of
  different summary levels.

  \cr 010 us \cr 020 region \cr 030 division \cr 040 state \cr 050 state,
  county \cr 060 state, county, county.subdivision \cr 140 state, county,
  tract \cr 150 state, county, tract, block.group \cr 160 state, place \cr
  250 american.indian.area \cr 320 state, msa \cr 340 state, csa \cr 350
  necta \cr 400 urban.area \cr 500 state, congressional.district \cr 610
  state, state.legislative.district.upper \cr 620 state,
  state.legislative.district.lower \cr 795 state, puma \cr 860 zip.code \cr
  950 state, school.district.elementary \cr 960 state,
  school.district.secondary \cr 970 state, school.district.unified \cr}

\item{combine.name}{Label for the aggregate geography when data are combined. The default value is \code{aggregate}.}

\item{print.levels}{Boolean that print levels generated by the \code{geo.make} function.}
}
\value{
Returns a combined \code{geo.make} object used to extract data.
}
\description{
Combine geographic objects created by the ACS package. Auxiliary function to combine geographic levels.
}
