% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_vars.R
\name{getvars}
\alias{getvars}
\title{Add together two numbers.}
\usage{
getvars(formula)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}

