% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_geo_level_data.R
\name{make_geo_level_data}
\alias{make_geo_level_data}
\title{Separate full raw data by geography}
\usage{
make_geo_level_data(source_data, codebook, geo, geoid_name)
}
\arguments{
\item{source_data}{Full source data}

\item{codebook}{ACS codebook}

\item{geo}{Geography name}

\item{geoid_name}{Name of the GEOID field}
}
\value{
A list of data frames
}
\description{
Separate full raw data by geography
}
