% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_year_geo_map.R
\name{make_year_geo_map}
\alias{make_year_geo_map}
\title{Create a data.frame map of year and geography}
\usage{
make_year_geo_map(years, geographies)
}
\arguments{
\item{years}{Vector of years}

\item{geographies}{Vector of geographies}
}
\value{
A data.frame with all combinations of years and geographies
}
\description{
Create a data.frame map of year and geography
}
