% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_codebook.R
\name{process_codebook}
\alias{process_codebook}
\title{Process ACS codebook}
\usage{
process_codebook(
  codebook_path,
  calculated_field_code,
  acs_years,
  step = c("source", "process")
)
}
\arguments{
\item{codebook_path}{Path to codebook CSV}

\item{calculated_field_code}{Code for calculated fields in the codebook}

\item{acs_years}{Years to pull data for}

\item{step}{One of 'source' or 'process', which determines whether source or derived fields are kept}
}
\value{
A data.table
}
\description{
Process ACS codebook
}
