% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_acs.R
\name{pull_acs}
\alias{pull_acs}
\title{Pull ACS data for a given geography, year, and set of variables}
\usage{
pull_acs(geo, acs_year, variables)
}
\arguments{
\item{geo}{Name of the geography to get ACS data for}

\item{acs_year}{Year to pull data for}

\item{variables}{Variables to pull}
}
\value{
A data.table
}
\description{
Pull ACS data for a given geography, year, and set of variables
}
