% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_data.R
\name{acsdata}
\alias{acsdata}
\title{Create a list containing ACS data files (\code{\link{acs}}).}
\usage{
acsdata(formula, level = "state", endyear = 2014, span = 5, us = "*",
  region = "*", division = "*", state = "WI", county = "*",
  county.subdivision = "*", place = "*", tract = "*", block.group = "*",
  msa = "*", csa = "*", necta = "*", urban.area = "*",
  congressional.district = "*", state.legislative.district.upper = "*",
  state.legislative.district.lower = "*", puma = "*", zip.code = "*",
  american.indian.area = "*", school.district.elementary = "*",
  school.district.secondary = "*", school.district.unified = "*",
  combine = FALSE, combine.name = "aggregate")
}
\arguments{
\item{formula}{A character or vector of characters containing formulas using
ACS variables. + - operators can be included. / defines a division.}

\item{level}{A character or vector of characters specifying the geographic
  level of the data. It may be necessary to specificy values to the
  corresponding levels. For instance, when \code{level = "county"}, you have
  to specify a state (e.g., \code{state = "WI"}, the default state in this
  package). You can also use a wildcard method (\code{state = "*"}) to
  include all the states. Below, you can see the required combinations of
  different summary levels.

  \cr 010 us \cr 020 region \cr 030 division \cr 040 state \cr 050 state,
  county \cr 060 state, county, county.subdivision \cr 140 state, county,
  tract \cr 150 state, county, tract, block.group \cr 160 state, place \cr
  250 american.indian.area \cr 320 state, msa \cr 340 state, csa \cr 350
  necta \cr 400 urban.area \cr 500 state, congressional.district \cr 610
  state, state.legislative.district.upper \cr 620 state,
  state.legislative.district.lower \cr 795 state, puma \cr 860 zip.code \cr
  950 state, school.district.elementary \cr 960 state,
  school.district.secondary \cr 970 state, school.district.unified \cr}

\item{endyear}{An integer (defaults to 2014) indicating the latest year of
the data in the survey.}

\item{span}{An integer indicating the span (in years) of the desired ACS data
(should be 1, 3,or 5), defaults to 5.}

\item{combine}{Whether the geographies are to be combined. If \code{combine =  TRUE}, lists should be used when specifying geographic levels (the corresponding level for the level specified). If the rest geographic levels has one element, the function will assume that level is equal for all the sub-levels. For example, if \code{state = "WI"}, and several counties were specified, the function assumes that all the counties are from WI.}

\item{combine.name}{Label for the aggregate geography when combining levels.
The default value is \code{aggregate}.}
}
\value{
Returns a list of ACS objects for different levels to be used with
  the \code{\link{sumacs}} function.
}
\description{
Because downloading the data to compute a proportion, ratio or
  aggregation can be slow, the \code{acsdata} function creates an object that  can be used with the \code{\link{sumacs}} function. Before using this
  function remember to define a key using the \code{\link{acs}} command
  \code{api.key.install(key="*")}.
}
\note{
Depending on the quality of the internet connection, number of
  variables and levels, getting the ACS data can be slow especially for the
  levels "county.subdivision", "block.group", and "tract" (it might take more
  than 30 minutes).
}
\examples{
api.key.install(key="*")
# get variables from formula
acsdata("(b16004_004 + b16004_026 + b16004_048 / b16004_001)", level = "county")

# get variables directly
acsdata(c("b16004_004", "b16004_026"), level = "county")

# combine levels
acsdata("(b16004_004 + b16004_026 + b16004_048 / b16004_001)",
  level = c("block.group"), state = list("WI"),
  county = list(1, 141),
  tract = list(950100, 11700),
  block.group = list(1:2, 1:2), combine = TRUE)
}

