% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acs_data.R
\name{acsdata}
\alias{acsdata}
\title{Create a list containing ACS data files.}
\usage{
acsdata(formula, level = "county", endyear = 2013, span = 5,
  state = "WI", county = "*", county_subdivision = "*", tract = "*",
  block_group = "*", congressional_district = "*",
  school_district_secondary = "*", school_district_elementary = "*")
}
\arguments{
\item{formula}{A character object or vector of characters containing formulas using ACS variables following the format.}

\item{level}{A character or vector of characters specifying the geographic level of the data. The levels included in this functions could be: "state", "county", "county.subdivision", "tract", "block.group", "congressional.district", "school.district.secondary", "school.district.elementary". The default value is "county".}

\item{endyear}{An integer (defaults to 2013) indicating the latest year of the data in the survey.}

\item{span}{An integer indicating the span (in years) of the desired ACS data (should be 1, 3,or 5), defaults to 5.}

\item{state}{Either the two-digit numeric FIPS code for the state, the two-letter postal abbreviation, or a character string to match in the state name (or wildcard "*" for all); setting state without other options corresponds to using census summary level 040, but it may be used in conjunction with other summary levels below. The default value is "WI".}

\item{county}{Either the numeric FIPS code (or wildcard "*" for all) for the county or a character string to match in the county name; setting state and county without other options corresponds to using census summary level 050, but they may be used in conjunction with other summary levels below. The default value is "*".}

\item{county_subdivision}{Either the numeric FIPS code (or wildcard "*" for all) for the county subdivision or a character string to match in the county subdivision name; setting state, county, and county.subdivision without other options corresponds to using census summary level 060.}

\item{tract}{A six digit numeric FIPS code (or wildcard "*" for all) for the census tract, including trailing zeroes; remove decimal points; leading zeroes may be omitted; see description; tract may be used with state and county for census summary levels 140, and with state, county, and block.group for summary level 150.}

\item{block_group}{The numeric FIPS code (or wildcard "*" for all) for the block.group.  block.group may be used with state, county, and tract for census summary levels 150. The default value is "*".}

\item{congressional_district}{A numeric code (or wildcard "*" for all) corresponding to the desired FIPS congressional district (e.g., state="WI" and congressional.district=1); setting state and congressional.district without other options corresponds to using census summary level 500, but they may be used in conjunction with county for summary level 510. The default value is "*".}

\item{school_district_secondary}{A numeric code (or wildcard "*" for all) corresponding to the desired FIPS state school district (secondary), or a character string to search for in the names of these districts; setting state and school.district.secondary without other options corresponds to using census summary level 960. The default value is "*".}

\item{school_district_elementary}{A numeric code (or wildcard "*" for all) corresponding to the desired FIPS state school district (elementary), or a character string to search for in the names of these districts; setting state and school.district.elementary without other options corresponds to using census summary level 950. The default value is "*".}
}
\value{
Returns a list of ACS objects for different levels to be used with the \code{\link{sumacs}} function.
}
\description{
Because downloading the data each time one wants to compute a proportion, ratio or aggregation can be time consuming, the \code{acsdata} function allow to save an object that can be used as input in the \code{\link{sumacs}} function. Before using this function remember to define a key using the \code{\link{acs}} command \code{api.key.install(key="*")}.
}
\note{
Depending on the quality of the internet connection, number of variables and level, getting the ACS data can be slow (it might take more than 30 minutes), especially for the levels "county.subdivision", "block.group", and "tract".
}
\examples{
api.key.install(key="*")
acsdata("(b16004_004 + b16004_026 + b16004_048 / b16004_001)", level = "county")
}

