% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_acs.R
\name{sumacs}
\alias{sumacs}
\title{Estimate proportions, ratios, aggregations and the respective margins
  of error (MOEs).}
\usage{
sumacs(formula, varname = NULL, method = NULL, level = "state",
  endyear = 2014, span = 5, conf.level = 0.9, one.zero = TRUE,
  trace = TRUE, data = NULL, format.out = "wide", file = NULL,
  print.levels = TRUE, us = "*", region = "*", division = "*",
  state = "WI", county = "*", county.subdivision = "*", place = "*",
  tract = "*", block.group = "*", msa = "*", csa = "*", necta = "*",
  urban.area = "*", congressional.district = "*",
  state.legislative.district.upper = "*",
  state.legislative.district.lower = "*", puma = "*", zip.code = "*",
  american.indian.area = "*", school.district.elementary = "*",
  school.district.secondary = "*", school.district.unified = "*",
  combine = FALSE, combine.name = "aggregate")
}
\arguments{
\item{formula}{A character or vector of characters containing formulas using
ACS variables or just variables. + - operators can be included. / defines a
division. When the formula contains "* 100", the final statistic and MOE is
multiply by 100.}

\item{varname}{A character or vector of characters containing the new
variables to be created. This vector must have same length as
\code{formula} and \code{method}, and it is not needed when getting only
variables.}

\item{method}{A character or vector of characters defining the type of
estimate expected: "proportion", "ratio", "aggregation", "variables". This
vector must have same length as \code{formula} and \code{varname}. It is
not needed when getting only variables.}

\item{level}{A character or vector of characters specifying the geographic
  level of the data. It may be necessary to specificy values to the
  corresponding levels. For instance, when \code{level = "county"}, you have
  to specify a state (e.g., \code{state = "WI"}, the default state in this
  package). You can also use a wildcard method (\code{state = "*"}) to
  include all the states. Below, you can see the required combinations of
  different summary levels.

  \cr 010 us \cr 020 region \cr 030 division \cr 040 state \cr 050 state,
  county \cr 060 state, county, county.subdivision \cr 140 state, county,
  tract \cr 150 state, county, tract, block.group \cr 160 state, place \cr
  250 american.indian.area \cr 320 state, msa \cr 340 state, csa \cr 350
  necta \cr 400 urban.area \cr 500 state, congressional.district \cr 610
  state, state.legislative.district.upper \cr 620 state,
  state.legislative.district.lower \cr 795 state, puma \cr 860 zip.code \cr
  950 state, school.district.elementary \cr 960 state,
  school.district.secondary \cr 970 state, school.district.unified \cr

When \code{combine == TRUE}, the geographic information should be in a list.}

\item{endyear}{An integer (defaults to 2014  ) indicating the latest year of
the data in the survey.}

\item{span}{An integer indicating the span (in years) of the desired ACS data
(should be 1, 3, or 5), defaults to 5.}

\item{conf.level}{Confidence level to estimate MOEs. The default value is
0.90.}

\item{one.zero}{Whether to include standard errors for only one zero-value
(max value) of columns or all. The default is TRUE.}

\item{trace}{Shows progress of the variable creation. The default is TRUE.}

\item{data}{Input data generated by the \code{\link{acsdata}} function.
Variables and levels must be the same as those specified in the
\code{sumacs} function.}

\item{format.out}{Format of the output: "wide" or "long". The default is
"wide".}

\item{file}{The resulting output is exported to a CSV file rather than to the R prompt. The file name must be specified as a character string.}

\item{print.levels}{Boolean that print levels generated by the \code{geo.make} function.}

\item{combine}{Whether the geographies are to be combined. If \code{combine = TRUE}, lists should be used when specifying geographic levels (the corresponding level for the level specified). If the rest geographic levels has one element, the function will assume that level is equal for all the sub-levels. For example, if \code{state = "WI"}, and several counties were specified, the function assumes that all the counties are from WI.}

\item{combine.name}{Label for the aggregate geography when combining levels. The default value is \code{aggregate}.}
}
\value{
Returns a \code{data.table/data.frame} object with the estimates and
  MOEs.
}
\description{
The \code{sumacs} function uses outputs from the
  \code{\link{acs}} package to compute proportions, ratios and aggregations
  based on text formulas, or simply extract variables. The function
  downloads the data and then estimate the formulas. If the function is used
  without specifying any \code{data}, remember to define a key using the
  \code{\link{acs}} command \code{api.key.install(key="*")}.
}
\details{
When the standard error of a proportion cannot be estimated, the
  "ratio" option is used. This adjustment is done row by row.
}
\note{
Depending on the quality of the internet connection, number of
  variables and levels, getting the ACS data can be slow, especially for the
  levels "county.subdivision", "block.group", and "tract" (it might take more than 30 minutes).  It is recommended to get the data using the function
  \code{\link{acsdata}} first, and then to use \code{sumacs}.
}
\examples{
# api.key.install(key="*")

# without combining
sumacs(formula = "(b16004_004 + b16004_026 + b16004_048 / b16004_001)",
 varname = "langspan0913", method = "prop")

# combining
sumacs("(b16004_004 + b16004_026 + b16004_048 / b16004_001)",
 varname = "test",
 method = "prop",
 level = c("block.group"),
 state = list("WI"),
 county = list(1, 141),
 tract = list(950100, 11700),
 block.group = list(1:2, 1:2),
 combine = TRUE)
}

